popayan_tourism/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.6
jwcrypto==1.5.4
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==1.7.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
